chore: refactor with absolute imports
